feat(js): Buffer read functionality (#2193)
* wip

* wip

* add buffer support

* feat(js): add buff support for json

* feat: add "Buffer" as supported read option
