+++ nodejs-polars/__tests__/io.test.ts
@@ -10,7 +10,8 @@ describe("read:csv", () => {
     const df = pl.readCSV(csvpath);
     expect(df.shape).toStrictEqual({height: 27, width: 4});
   });
-  it.only("can read from a relative file", () => {
+
+  it("can read from a relative file", () => {
     const df = pl.readCSV("../examples/aggregate_multiple_files_in_chunks/datasets/foods1.csv");
     expect(df.shape).toStrictEqual({height: 27, width: 4});
   });
@@ -22,6 +23,16 @@ describe("read:csv", () => {
     const csvString = "foo,bar,baz\n1,2,3\n4,5,6\n";
     const df = pl.readCSV(csvString);
     expect(df.toCSV()).toEqual(csvString);
+  });
+  it("can read from a csv buffer", () => {
+    const csvBuffer = Buffer.from("foo,bar,baz\n1,2,3\n4,5,6\n", "utf-8");
+    const df = pl.readCSV(csvBuffer);
+    expect(df.toCSV()).toEqual(csvBuffer.toString());
+  });
+  it("can read from a csv buffer with options", () => {
+    const csvBuffer = Buffer.from("foo,bar,baz\n1,2,3\n4,5,6\n", "utf-8");
+    const df = pl.readCSV(csvBuffer, {hasHeader: true, batchSize: 10});
+    expect(df.toCSV()).toEqual(csvBuffer.toString());
   });
   it("can parse datetimes", () => {
     const csv = `timestamp,open,high
@@ -50,7 +61,7 @@ describe("read:json", () => {
     expect(df.shape).toStrictEqual({height: 27, width: 4});
   });
   it("can specify read options", () => {
-    const df = pl.readJSON({file: jsonpath, batchSize: 10, inferSchemaLength: 100});
+    const df = pl.readJSON(jsonpath, {batchSize: 10, inferSchemaLength: 100});
     expect(df.shape).toStrictEqual({height: 27, width: 4});
   });
   it("can read from a json string", () => {
@@ -59,8 +70,8 @@ describe("read:json", () => {
     expect(df.toJSON().replace("\n", "")).toEqual(jsonString);
   });
 });
-describe.skip("scan", () => {
-  describe.skip("csv", () => {
+describe("scan", () => {
+  describe("csv", () => {
     it("can lazy load (scan) from a csv file", () => {
       const df = pl.scanCSV(csvpath).collectSync();
       expect(df.shape).toStrictEqual({height: 27, width: 4});
